# Optimize compiler.(*engine).NewModuleEngine