Before:
```sh
❯ go test -benchmem -memprofile memprofile.out -cpuprofile profile.out -run=^$ -bench ^BenchmarkInitialization/compiler$ github.com/tetratelabs/wazero/internal/integration_test/bench
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
BenchmarkInitialization/compiler-10                64510             17417 ns/op          145950 B/op        128 allocs/op
PASS
ok      github.com/tetratelabs/wazero/internal/integration_test/bench   1.600s
```
<img width="1340" alt="Screenshot 2022-11-05 at 14 06 24" src="https://user-images.githubusercontent.com/1851985/200141469-a57e30a8-6732-4f70-acf6-f4f72e41bdd1.png">

After:
```sh
❯ go test -benchmem -memprofile memprofile.out -cpuprofile profile.out -run=^$ -bench ^BenchmarkInitialization/compiler$ github.com/tetratelabs/wazero/internal/integration_test/bench
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
BenchmarkInitialization/compiler-10                59250             17296 ns/op          145950 B/op        128 allocs/op
PASS
ok      github.com/tetratelabs/wazero/internal/integration_test/bench   1.493s
```
<img width="1354" alt="Screenshot 2022-11-05 at 14 15 46" src="https://user-images.githubusercontent.com/1851985/200141634-185e28d7-5345-4742-bb12-3f578854e855.png">

Another pretty minor win. Includes commits from #841, if that one can get merged first this will be a clean diff.