@ckaznocha 2022-11-05
Not sure why the ci timed out, it passes when I run it on my laptop
```sh
❯ make spectest
ok      github.com/tetratelabs/wazero/internal/integration_test/spectest/v1     2.197s
ok      github.com/tetratelabs/wazero/internal/integration_test/spectest/v2     3.307s
```