@mathetake 2022-11-06
freebsd tends to timeout occationally, so you can just ignore the ci error on that one