@ckaznocha 2022-11-06
I regret trying to stack changes now. If this one is acceptable, it can just get merged and #841 should auto close or it can be manually closed.